implemented max queue size + SenderBuilder (it was time)
