+++ src/header.rs
@@ -13,6 +13,7 @@
 //     print()
 // ```
 // See: https://en.wikipedia.org/wiki/Hamming_code#General_algorithm
+const P0: u32 = 0b11_1111_1111_1111_1111_1111_1111; // just a regular parity chech (not Hamming!)
 const P1: u32 = 0b10_1010_1010_1010_1101_0101_1011;
 const P2: u32 = 0b11_0011_0011_0011_0110_0110_1101;
 const P3: u32 = 0b11_1100_0011_1100_0111_1000_1110;
@@ -22,7 +23,6 @@ const P5: u32 = 0b11_1111_1111_1111_1000_0000_0000;
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct Header {
     len: u32,
-    is_legacy: bool,
 }
 
 impl Header {
@@ -36,14 +36,7 @@ impl Header {
 
         Header {
             len,
-            is_legacy: false,
         }
-    }
-
-    fn new_legacy(len: u32) -> Header {
-        let mut header = Header::new(len);
-        header.is_legacy = true;
-        header
     }
 
     pub fn len(&self) -> u32 {
@@ -51,15 +44,10 @@ impl Header {
     }
 
     pub fn encode(&self) -> [u8; 4] {
-        assert!(!self.is_legacy, "encoding in legacy mode is deprecated");
-
         let mut encoded = self.len;
 
-        // set first flag bit (new stuff):
-        encoded |= 1 << 26;
-
         // parities:
-        // len |= (len.count_ones() & 0b1) << 26; // future...
+        encoded |= ((self.len & P0).count_ones() & 0b1) << 26;
         encoded |= ((self.len & P1).count_ones() & 0b1) << 27;
         encoded |= ((self.len & P2).count_ones() & 0b1) << 28;
         encoded |= ((self.len & P3).count_ones() & 0b1) << 29;
@@ -72,69 +60,63 @@ impl Header {
     pub fn decode(header: [u8; 4]) -> Header {
         let encoded = u32::from_be_bytes(header);
 
-        if encoded & (1 << 26) != 0 {
-            // first flag set: new stuff
-            // check parities...
+        // check parities...
 
-            fn xor(a: bool, b: bool) -> bool {
-                a && b || !a && !b
-            }
-
-            // assert!(
-            //     xor(
-            //         len.count_ones() as u8 & 0b1 != 0,
-            //         len & (1 << 26) != 0
-            //     ),
-            //     "parity 1 wrong: {:?}",
-            //     header
-            // );
-            assert!(
-                xor(
-                    (encoded & P1).count_ones() as u8 & 0b1 != 0,
-                    encoded & (1 << 27) != 0
-                ),
-                "parity 1 wrong: {:?}",
-                header
-            );
-            assert!(
-                xor(
-                    (encoded & P2).count_ones() as u8 & 0b1 != 0,
-                    encoded & (1 << 28) != 0
-                ),
-                "parity 2 wrong: {:?}",
-                header
-            );
-            assert!(
-                xor(
-                    (encoded & P3).count_ones() as u8 & 0b1 != 0,
-                    encoded & (1 << 29) != 0
-                ),
-                "parity 3 wrong: {:?}",
-                header
-            );
-            assert!(
-                xor(
-                    (encoded & P4).count_ones() as u8 & 0b1 != 0,
-                    encoded & (1 << 30) != 0
-                ),
-                "parity 4 wrong: {:?}",
-                header
-            );
-            assert!(
-                xor(
-                    (encoded & P5).count_ones() as u8 & 0b1 != 0,
-                    encoded & (1 << 31) != 0
-                ),
-                "parity 5 wrong: {:?}",
-                header
-            );
-
-            // clear flags
-            Header::new(encoded & 0x03_FF_FF_FFu32)
-        } else {
-            // last bit not set: old stuff
-            Header::new_legacy(encoded)
+        fn xor(a: bool, b: bool) -> bool {
+            a && b || !a && !b
         }
+
+        assert!(
+            xor(
+                (encoded & P0).count_ones() as u8 & 0b1 != 0,
+                encoded & (1 << 26) != 0
+            ),
+            "parity 0 wrong: {:?}",
+            header
+        );
+        assert!(
+            xor(
+                (encoded & P1).count_ones() as u8 & 0b1 != 0,
+                encoded & (1 << 27) != 0
+            ),
+            "parity 1 wrong: {:?}",
+            header
+        );
+        assert!(
+            xor(
+                (encoded & P2).count_ones() as u8 & 0b1 != 0,
+                encoded & (1 << 28) != 0
+            ),
+            "parity 2 wrong: {:?}",
+            header
+        );
+        assert!(
+            xor(
+                (encoded & P3).count_ones() as u8 & 0b1 != 0,
+                encoded & (1 << 29) != 0
+            ),
+            "parity 3 wrong: {:?}",
+            header
+        );
+        assert!(
+            xor(
+                (encoded & P4).count_ones() as u8 & 0b1 != 0,
+                encoded & (1 << 30) != 0
+            ),
+            "parity 4 wrong: {:?}",
+            header
+        );
+        assert!(
+            xor(
+                (encoded & P5).count_ones() as u8 & 0b1 != 0,
+                encoded & (1 << 31) != 0
+            ),
+            "parity 5 wrong: {:?}",
+            header
+        );
+
+        // clear flags
+        Header::new(encoded & 0x03_FF_FF_FFu32)
     }
 }
 
@@ -155,17 +137,6 @@ mod test {
         for (_i, header) in lengths.enumerate() {
             // println!("{}", i);
             assert_eq!(header, Header::decode(header.encode()));
-        }
-    }
-
-    #[test]
-    fn encode_length_legacy_ok() {
-        let lengths = lots_of_lengths();
-
-        for (_i, header) in lengths.enumerate() {
-            // println!("{}", i);
-            let len = header.len();
-            assert_eq!(len, Header::decode(u32::to_be_bytes(len)).len());
         }
     }
